+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This changelog references the relevant changes done between versions.
 To get the diff for a specific change, go to https://github.com/LIN3S/FrontFoundation/commit/XXX where XXX is the change hash 
 To get the diff between two versions, go to https://github.com/LIN3S/FrontFoundation/compare/v0.5.0...v0.6.0
 
+* 0.11.4
+    * Fixed focus/blur bug on the `FormSelect` react atom.
 * 0.11.3
 * 0.11.2
 * 0.11.1